ntkn.bcknd/requirements.txt
slugify
Django
Unidecode==0.04.17
defusedxml==0.4.1
django-allauth==0.19.1
django-appconf==1.0.1
django-compressor==1.5
django-localflavor==1.1
django-rest-auth==0.4.0
djangorestframework==3.1.1
oauthlib==0.7.2
python3-openid==3.0.5
regex==2015.03.18
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
django-imagekit
